feat: 랜딩 페이지 인터랙션 적용
