+++ src/pages/landing/LandingPage.css.ts
@@ -2,7 +2,22 @@ import { style } from '@vanilla-extract/css';
 
 import { colorVars } from '@styles/tokensV2/color.css';
 import { fontVars } from '@styles/tokensV2/font.css';
+import {
+  interaction,
+  type InteractionSpec,
+} from '@styles/tokensV2/interaction/interaction.utils';
 import { unitVars } from '@styles/tokensV2/unit.css';
+
+/** afterDelay 2s → stateChange (크로스페이드) */
+const landingBgTransitionInteraction = {
+  trigger: 'afterDelay',
+  action: 'stateChange',
+  duration: 'slower',
+  easing: 'bezier.inout',
+  property: 'opacity',
+} as const satisfies InteractionSpec;
+
+const landingBgTransition = interaction(landingBgTransitionInteraction);
 
 export const page = style({
   position: 'relative',
@@ -20,7 +35,7 @@ export const backgroundImage = style({
   position: 'absolute',
   zIndex: 0,
   inset: 0,
-  transition: 'opacity 600ms ease',
+  transition: landingBgTransition,
   opacity: 0,
   objectFit: 'cover',
   // 모든 배경 이미지를 겹쳐두고 현재 이미지만 opacity 1로 올려 dissolve(크로스페이드). 리마운트 없음.
